+++ sdk/src/androidTest/java/ly/count/android/sdk/RemoteConfigVariantControlTests.java
@@ -181,7 +181,7 @@ public class RemoteConfigVariantControlTests {
 
     @Test
     public void testNormalFlow() {
-        CountlyConfig config = TestUtils.createVariantConfig(createIRGForSpecificResponse("{\"key\":[{\"name\":\"variant\"}]}"));
+        CountlyConfig config = TestUtils.createIRGeneratorConfig(createIRGForSpecificResponse("{\"key\":[{\"name\":\"variant\"}]}"));
         Countly countly = new Countly().init(config);
 
         // Developer did not provide a callback
@@ -204,7 +204,7 @@ public class RemoteConfigVariantControlTests {
      */
     @Test
     public void testNullVariant() {
-        CountlyConfig config = TestUtils.createVariantConfig(createIRGForSpecificResponse("{\"key\":[{\"name\":null}]}"));
+        CountlyConfig config = TestUtils.createIRGeneratorConfig(createIRGForSpecificResponse("{\"key\":[{\"name\":null}]}"));
         Countly countly = new Countly().init(config);
 
         // Developer did not provide a callback
@@ -221,7 +221,7 @@ public class RemoteConfigVariantControlTests {
      */
     @Test
     public void testFilteringWrongKeys() {
-        CountlyConfig config = TestUtils.createVariantConfig(createIRGForSpecificResponse("{\"key\":[{\"noname\":\"variant1\"},{\"name\":\"variant2\"}]}"));
+        CountlyConfig config = TestUtils.createIRGeneratorConfig(createIRGForSpecificResponse("{\"key\":[{\"noname\":\"variant1\"},{\"name\":\"variant2\"}]}"));
         Countly countly = new Countly().init(config);
 
         // Developer did not provide a callback
@@ -255,7 +255,7 @@ public class RemoteConfigVariantControlTests {
 
     @Test
     public void variantGetters_preDownload() {
-        CountlyConfig config = TestUtils.createVariantConfig(null);
+        CountlyConfig config = TestUtils.createIRGeneratorConfig(null);
         Countly countly = new Countly().init(config);
 
         //should return empty map of values