+++ sdk/src/androidTest/java/ly/count/android/sdk/scUP_UserProfileTests.java
@@ -22,13 +22,13 @@ public class scUP_UserProfileTests {
 
     @Before
     public void setUp() {
-        TestUtils.getCountyStore().clear();
+        TestUtils.getCountlyStore().clear();
         Countly.sharedInstance().halt();
     }
 
     @After
     public void tearDown() {
-        TestUtils.getCountyStore().clear();
+        TestUtils.getCountlyStore().clear();
         Countly.sharedInstance().halt();
     }