+++ sdk/src/main/java/ly/count/android/sdk/ConfigurationProvider.java
@@ -5,5 +5,17 @@ interface ConfigurationProvider {
 
     boolean getTrackingEnabled();
 
+    boolean getSessionTrackingEnabled();
+
+    boolean getViewTrackingEnabled();
+
+    boolean getCustomEventTrackingEnabled();
+
+    boolean getContentZoneEnabled();
+
     boolean getCrashReportingEnabled();
+
+    boolean getLocationTrackingEnabled();
+
+    boolean getRefreshContentZoneEnabled();
 }