+++ sdk/src/main/java/ly/count/android/sdk/CountlyConfig.java
@@ -79,7 +79,7 @@ public class CountlyConfig {
     protected String appKey = null;
 
     /**
-     * unique ID for the device the app is running on; note that null in deviceID means that Countly will fall back to OpenUDID, then, if it's not available, to Google Advertising ID.
+     * unique ID for the device the app is running on; note that null in deviceID means that Countly will fall back to UUID.
      */
     protected String deviceID = null;
 
@@ -201,6 +201,7 @@ public class CountlyConfig {
 
     // Requests older than this value in hours would be dropped (0 means this feature is disabled)
     int dropAgeHours = 0;
+    String sdkBehaviorSettings;
 
     /**
      * THIS VARIABLE SHOULD NOT BE USED
@@ -289,22 +290,12 @@ public class CountlyConfig {
     }
 
     /**
-     * unique ID for the device the app is running on; note that null in deviceID means that Countly will fall back to OpenUDID, then, if it's not available, to Google Advertising ID.
+     * unique ID for the device the app is running on; note that null in deviceID means that Countly will fall back to UUID.
      *
      * @return Returns the same config object for convenient linking
      */
     public synchronized CountlyConfig setDeviceId(String deviceID) {
         this.deviceID = deviceID;
-        return this;
-    }
-
-    /**
-     * enum value specifying which device ID generation strategy Countly should use: OpenUDID or Google Advertising ID.
-     *
-     * @return Returns the same config object for convenient linking
-     * @deprecated this call should not be used anymore as it does not have any purpose anymore
-     */
-    public synchronized CountlyConfig setIdMode(DeviceIdType idMode) {
         return this;
     }
 
@@ -992,7 +983,6 @@ public class CountlyConfig {
         return this;
     }
 
-    
     /**
      * This is an experimental feature and it can have breaking changes
      *
@@ -1005,9 +995,20 @@ public class CountlyConfig {
     public synchronized CountlyConfig enableServerConfiguration() {
         return this;
     }
-    
+
     protected synchronized CountlyConfig disableHealthCheck() {
         healthCheckEnabled = false;
+        return this;
+    }
+
+    /**
+     * Set the server configuration to be set while initializing the SDK
+     *
+     * @param sdkBehaviorSettings The server configuration to be set
+     * @return Returns the same config object for convenient linking
+     */
+    public synchronized CountlyConfig setSDKBehaviorSettings(String sdkBehaviorSettings) {
+        this.sdkBehaviorSettings = sdkBehaviorSettings;
         return this;
     }