+++ sdk/src/main/java/ly/count/android/sdk/CountlyTimer.java
@@ -30,6 +30,13 @@ class CountlyTimer {
         }
     }
 
+    /**
+     * Start a timer with the given delay
+     *
+     * @param timerDelay in seconds
+     * @param runnable to run
+     * @param L logger
+     */
     protected void startTimer(long timerDelay, @NonNull Runnable runnable, @NonNull ModuleLog L) {
         startTimer(timerDelay, 0, runnable, L);
     }