+++ sdk/src/main/java/ly/count/android/sdk/ModuleSessions.java
@@ -46,6 +46,10 @@ public class ModuleSessions extends ModuleBase {
             return;
         }
 
+        if (!configProvider.getSessionTrackingEnabled()) {
+            return;
+        }
+
         if (sessionIsRunning()) {
             L.w("[ModuleSessions] A session is already running, this 'beginSessionInternal' will be ignored");
             healthTracker.logSessionStartedWhileRunning();
@@ -70,6 +74,10 @@ public class ModuleSessions extends ModuleBase {
             return;
         }
 
+        if (!configProvider.getSessionTrackingEnabled()) {
+            return;
+        }
+
         if (!sessionIsRunning()) {
             L.w("[ModuleSessions] No session is running, this 'updateSessionInternal' will be ignored");
             healthTracker.logSessionUpdatedWhileNotRunning();
@@ -85,6 +93,10 @@ public class ModuleSessions extends ModuleBase {
         L.d("[ModuleSessions] endSessionInternal, checkConsent:[" + checkConsent + "]");
 
         if (checkConsent && !consentProvider.getConsent(Countly.CountlyFeatureNames.sessions)) {
+            return;
+        }
+
+        if (!configProvider.getSessionTrackingEnabled()) {
             return;
         }