+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## 25.4.0
+* ! Minor breaking change ! Removed Secure.ANDROID_ID usage in device id generation. The SDK now exclusively uses random UUIDs for device id generation.
+
 * Added a Content feature method "refreshContentZone" that does a manual refresh.
+* Extended server configuration capabilities of the SDK.
+* Added a config method to provide server config in the initialization "setSDKBehaviorSettings(String)".
 * Added a new interface "CountlyNotificationButtonURLHandler" to allow custom handling of URLs when notification buttons are clicked. Could be set by "CountlyConfigPush.setNotificationButtonURLHandler"
+
+* Mitigated an issue that caused PN message data collision if two message with same ID was received.
+
+* Removed the deprecated function "CountlyConfig.setIdMode(idMode)"
 
 * Deprecated the experimental configuration function enableServerConfiguration. It is now enabled by default and can be controlled directly from the server.